+++ .github/workflows/build.yml
@@ -17,7 +17,7 @@ jobs:
         java: [ 8, 11 ]
         # WARN: build.sbt depends on this key path, as scalaVersion and
         # crossScalaVersions is determined from it
-        scala: [ 2.11.12, 2.12.12, 2.13.3 ]
+        scala: [ 2.12.12, 2.13.3 ]
     
     env:
       CI: true
