+++ monix-catnap/jvm/src/main/scala/monix/catnap/internal/FutureLiftForPlatform.scala
