+++ monix-catnap/jvm/src/test/scala/monix/catnap/FutureLiftJava8Suite.scala
