+++ monix-execution/jvm/src/main/scala/monix/execution/internal/CancelableFutureForPlatform.scala
