+++ monix-execution/jvm/src/main/scala/monix/execution/internal/FutureUtilsForPlatform.scala
