+++ monix-execution/jvm/src/main/scala/monix/execution/internal/forkJoin/DynamicWorkerThreadFactory.scala
@@ -17,9 +17,12 @@
 
 package monix.execution.internal.forkJoin
 
-import java.util.concurrent.ThreadFactory
+import java.util.concurrent.ForkJoinPool.{ForkJoinWorkerThreadFactory, ManagedBlocker}
+import java.util.concurrent.{ForkJoinPool, ForkJoinWorkerThread, ThreadFactory}
+
 import monix.execution.atomic.AtomicInt
 import monix.execution.internal.forkJoin.DynamicWorkerThreadFactory.EmptyBlockContext
+
 import scala.annotation.tailrec
 import scala.concurrent.{BlockContext, CanAwait}
