+++ monix-execution/jvm/src/main/scala/monix/execution/internal/forkJoin/StandardWorkerThreadFactory.scala
@@ -17,7 +17,8 @@
 
 package monix.execution.internal.forkJoin
 
-import java.util.concurrent.ThreadFactory
+import java.util.concurrent.ForkJoinPool.ForkJoinWorkerThreadFactory
+import java.util.concurrent.{ForkJoinPool, ForkJoinWorkerThread, ThreadFactory}
 
 private[monix] final class StandardWorkerThreadFactory(
   prefix: String,
