+++ monix-execution/shared/src/main/scala/monix/execution/misc/InlineMacros.scala
@@ -17,7 +17,6 @@
 
 package monix.execution.misc
 
-import monix.execution.misc.compat.setOrig
 import scala.reflect.macros.whitebox
 
 trait InlineMacros {
@@ -49,7 +48,7 @@ trait InlineMacros {
           value
         case tt: TypeTree if tt.original != null =>
           //super.transform(TypeTree().setOriginal(transform(tt.original)))
-          super.transform(setOrig(c)(TypeTree(), transform(tt.original)))
+          super.transform(c.universe.internal.setOriginal(TypeTree(), transform(tt.original)))
         case _ =>
           super.transform(tree)
       }
