--- monix-execution/shared/src/main/scala_2.11/monix/execution/misc/compat.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2014-2020 by The Monix Project Developers.
- * See the project homepage at: https://monix.io
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package monix.execution.misc
-
-private[execution] object compat {
-  type Context = scala.reflect.macros.whitebox.Context
-
-  def freshTermName[C <: Context](c: C)(s: String) =
-    c.universe.TermName(c.freshName(s))
-
-  def setOrig[C <: Context](c: C)(tt: c.universe.TypeTree, t: c.Tree) =
-    c.universe.internal.setOriginal(tt, t)
-}
